+++ src/app.ts
@@ -34,6 +34,7 @@ import * as postboxItemAPI from "./routes/postbox";
 import * as topUpsAPI from "./routes/topUps";
 import * as instantCreditTransferAPI from "./routes/instantCreditTransfer";
 import * as accountOpeningRequestAPI from "./routes/accountOpeningRequest";
+import * as intraCustomerTransferApi from "./routes/intraCustomerTransfer";
 
 import { migrate } from "./db";
 
@@ -453,7 +454,7 @@ router.post(
 
 router.post(
   "/persons/:person_id/accounts/:account_id/transactions/intra_customer_transfer",
-  safeRequestHandler(transactionsAPI.createIntraCustomerTransfer)
+  safeRequestHandler(intraCustomerTransferApi.createIntraCustomerTransfer)
 );
 
 // STANDING ORDERS
